chore: optimize

+++ app/src/main/java/cn/morfans/chenjunyu19/mediavolume/Media2Service.java
@@ -2,6 +2,6 @@ package cn.morfans.chenjunyu19.mediavolume;
 
 public class Media2Service extends MediaTileService {
     public Media2Service() {
-        super((double) 2 / 3);
+        super(2f / 3f);
     }
 }
