+++ app/src/main/java/cn/morfans/chenjunyu19/mediavolume/MediaHalfService.java
@@ -2,6 +2,6 @@ package cn.morfans.chenjunyu19.mediavolume;
 
 public class MediaHalfService extends MediaTileService {
     public MediaHalfService() {
-        super((double) 1 / 2);
+        super(1f / 2f);
     }
 }
