+++ app/src/main/java/cn/morfans/chenjunyu19/mediavolume/MediaMaxService.java
@@ -2,6 +2,6 @@ package cn.morfans.chenjunyu19.mediavolume;
 
 public class MediaMaxService extends MediaTileService {
     public MediaMaxService() {
-        super((double) 1);
+        super(1f);
     }
 }
